+++ R/as_index.R
@@ -14,9 +14,9 @@
 #' The data frame method for `as_index()` is best understood as reversing
 #' the effect of [`as.data.frame()`][as.data.frame.piar_index] on an
 #' index object. It constructs a matrix by taking the levels of
-#' `x[[cols[1]]]` as columns and the levels of `x[[cols[2]]]` as rows
+#' `x[[1]]` as columns and the levels of `x[[1]]` as rows
 #' (coercing to a factor if necessary). It then populates this matrix with the
-#' corresponding values in `x[[cols[3]]]`, and uses the matrix method for
+#' corresponding values in `x[[3]]`, and uses the matrix method for
 #' `as_index()`.
 #'
 #' If `x` is a period-over-period index then it is returned unchanged when
@@ -28,8 +28,8 @@
 #' @param chainable Are the index values in `x` period-over-period
 #' indexes, suitable for a chained calculation (the default)? This should be
 #' `FALSE` when `x` is a fixed-base (direct) index.
-#' @param cols A vector giving the positions/names of the period, level, and
-#' value columns in `x`. The default assumes that the first column
+#' @param cols  Deprecated. A vector giving the positions/names of the period,
+#' level, and value columns in `x`. The default assumes that the first column
 #' contains time periods, the second contains levels, and the third contains
 #' index values.
 #' @param ... Further arguments passed to or used by methods.
@@ -91,8 +91,18 @@ as_index.matrix <- function(x, chainable = TRUE, ...) {
 
 #' @rdname as_index
 #' @export
-as_index.data.frame <- function(x, cols = 1:3, chainable = TRUE, ...) {
-  x <- x[cols]
+as_index.data.frame <- function(x, cols = NULL, chainable = TRUE, ...) {
+  if (length(x) < 3L) {
+    stop(
+      "'x' must have a column of time periods, index levels, and index values"
+    )
+  }
+  if (!is.null(cols)) {
+    warning("'cols' is deprecated and will be removed in a future version")
+    x <- x[cols]
+  } else {
+    x <- x[1:3]
+  }
   x[1:2] <- lapply(x[1:2], as.factor)
   time <- levels(x[[1L]])
   levels <- levels(x[[2L]])